# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/service-areas": {
            "get": {
                "description": "gets all service-areas in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "get all service-areas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/dto.serviceAreasResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "creates a new service-area",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "create service-area",
                "parameters": [
                    {
                        "description": "Add service-area",
                        "name": "service-area",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BodyCreateServiceArea"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ServiceAreaResponse"
                        }
                    }
                }
            }
        },
        "/api/service-areas/{id}": {
            "get": {
                "description": "gets a service-area from the system by its ID",
                "produces": [
                    "application/json"
                ],
                "summary": "get service-area",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Service-area id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ServiceAreaResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Area": {
            "type": "object",
            "properties": {
                "coordinates": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "array",
                            "items": {
                                "type": "number"
                            }
                        }
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "dto.BodyCreateServiceArea": {
            "type": "object",
            "properties": {
                "area": {
                    "$ref": "#/definitions/domain.Area"
                },
                "id": {
                    "type": "integer"
                },
                "identifier": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.ServiceAreaResponse": {
            "type": "object",
            "properties": {
                "area": {
                    "$ref": "#/definitions/domain.Area"
                },
                "id": {
                    "type": "integer"
                },
                "identifier": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.serviceAreasResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "identifier": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}